+++ addons/l10n_pl/demo/demo_company.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <record id="partner_demo_company_pl" model="res.partner">
-        <field name="name">PL Company</field>
+        <field name="name">My Polish Company</field>
         <field name="vat">PL5795955811</field>
         <field name="street"/>
         <field name="city">Racibórz</field>
@@ -14,7 +14,7 @@
     </record>
 
     <record id="demo_company_pl" model="res.company">
-        <field name="name">PL Company</field>
+        <field name="name">My Polish Company</field>
         <field name="partner_id" ref="partner_demo_company_pl"/>
     </record>
 
@@ -24,12 +24,6 @@
 
     <function model="res.users" name="write">
         <value eval="[ref('base.user_root'), ref('base.user_admin'), ref('base.user_demo')]"/>
-        <value eval="{'company_ids': [(4, ref('l10n_pl.demo_company_pl'))]}"/>
-    </function>
-
-    <function model="account.chart.template" name="try_loading">
-        <value eval="[]"/>
-        <value>pl</value>
-        <value model="res.company" eval="obj().env.ref('l10n_pl.demo_company_pl')"/>
+        <value eval="{'company_ids': [(4, ref('l10n_pl.demo_company_pl'))], 'tz': 'Europe/Warsaw'}"/>
     </function>
 </odoo>
