+++ addons/l10n_pl_account/i18n/l10n_pl.pot
