+++ addons/l10n_pl_account/models/res_company.py
