+++ addons/l10n_pl_account/views/product_views.xml
