+++ addons/l10n_pl_account/views/res_config_settings_views.xml
